launch meshcat visualizer node when using Simple
